# WIP: speed up Period creation