@jreback 2018-11-05
so we should benchmark a list of Periods created with an offset and from a string. not sure if these are benchmarked now.