@jreback 2018-12-03
@qwhelan can you merge master. pls retest the asv. does this show any significant speedups?