@qwhelan 2018-12-04
@jreback I believe the meat of the improvements was merged via https://github.com/pandas-dev/pandas/pull/23589, so I'm not seeing any significant changes when I rebase on master and re-run